+++ api/core/src/modules/identity/identity.module.ts
@@ -1,10 +1,11 @@
 import { Module } from "@nestjs/common";
-import { IdentityController } from "./identity.controller";
+import { AuthController, IdentityController, ProfileController } from "./identity.controller";
+import { IdentityRepository } from "./identity.repository";
 import { IdentityService } from "./identity.service";
 
 @Module({
-  controllers: [IdentityController],
-  providers: [IdentityService],
+  controllers: [IdentityController, AuthController, ProfileController],
+  providers: [IdentityRepository, IdentityService],
   exports: [IdentityService]
 })
 export class IdentityModule {}